requirements.txt
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
openai-whisper
numpy
pyobjc-framework-Quartz; sys_platform == 'darwin'
pyobjc-framework-Cocoa; sys_platform == 'darwin'
pynput; sys_platform != 'darwin'
pyautogui; sys_platform == 'win32'
pywin32; sys_platform == 'win32'
